+++ test/services/blob/blobservice-tests.js
@@ -1523,7 +1523,7 @@ describe('BlobService', function () {
         assert.strictEqual(parsedUrl.port, '80');
         assert.strictEqual(parsedUrl.hostname, 'host.com');
         assert.strictEqual(parsedUrl.pathname, '/' + containerName + '/' + blobName);
-        assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2017-07-29&sr=b&sig=o1TR6AHtC7jKDuMq6Y9bqOtCxTBiokFRHtKgv9VEFpQ%3D');
+        assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2018-03-28&sr=b&sig=j6Ubb%2Flpyt9cd55dlfwoEGOog9a%2FyDYWhB8UscVCTo4%3D');
 
         blobUrl = blobServiceassert.getUrl(containerName, blobName, sasToken, false, '2016-10-11T11:03:40Z');
 
@@ -1532,7 +1532,7 @@ describe('BlobService', function () {
         assert.strictEqual(parsedUrl.port, '80');
         assert.strictEqual(parsedUrl.hostname, 'host-secondary.com');
         assert.strictEqual(parsedUrl.pathname, '/' + containerName + '/' + blobName);
-        assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2017-07-29&sr=b&sig=o1TR6AHtC7jKDuMq6Y9bqOtCxTBiokFRHtKgv9VEFpQ%3D&snapshot=2016-10-11T11%3A03%3A40Z');
+        assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2018-03-28&sr=b&sig=j6Ubb%2Flpyt9cd55dlfwoEGOog9a%2FyDYWhB8UscVCTo4%3D&snapshot=2016-10-11T11%3A03%3A40Z');
 
         done();
       });
@@ -1613,7 +1613,7 @@ describe('BlobService', function () {
       assert.equal(sasQueryString[QueryStringConstants.SIGNED_PERMISSIONS], BlobUtilities.SharedAccessPermissions.READ);
       assert.equal(sasQueryString[QueryStringConstants.SIGNED_PROTOCOL], 'https');
       assert.equal(sasQueryString[QueryStringConstants.SIGNED_VERSION], HeaderConstants.TARGET_STORAGE_VERSION);
-      assert.equal(sasQueryString[QueryStringConstants.SIGNATURE], 'b6Do7ByS97s+IVkusKmJ56h7YJQh0FNMmX9h4ps/HcE=');
+      assert.equal(sasQueryString[QueryStringConstants.SIGNATURE], '8rgGI044z+1BGRrVKaxO+u5y/pqjBSPVqC3QDEJvTUA=');
 
       done();
     });
@@ -2077,6 +2077,62 @@ describe('BlobService', function () {
     assert.equal(blobService.host.primaryHost, 'http://127.0.0.1:10000/devstoreaccount1');
 
     done();
+  });
+
+  describe('StaticWebsite', function () {
+    it('should be able to enable static website', function (done) {
+      var prop = {
+        StaticWebsite: {
+          Enabled: true,
+          IndexDocument: 'index.htm',
+          ErrorDocument404Path: 'error/index.htm'
+        }
+      };
+
+      blobService.setServiceProperties(prop, function (err) {
+        assert.equal(err, null);
+
+        blobService.getServiceProperties(function (err, res) {
+          assert.equal(res.StaticWebsite.Enabled, true);
+          assert.equal(res.StaticWebsite.IndexDocument, prop.StaticWebsite.IndexDocument);
+          assert.equal(res.StaticWebsite.ErrorDocument404Path, prop.StaticWebsite.ErrorDocument404Path);
+          done();
+        });
+      });
+    });
+
+    it('should be able to disable static website', function (done) {
+      var prop = {
+        StaticWebsite: {
+          Enabled: false
+        }
+      };
+
+      blobService.setServiceProperties(prop, function (err) {
+        assert.equal(err, null);
+
+        blobService.getServiceProperties(function (err, res) {
+          assert.equal(res.StaticWebsite.Enabled, false);
+          assert.equal(res.StaticWebsite.IndexDocument, null);
+          assert.equal(res.StaticWebsite.ErrorDocument404Path, null);
+          done();
+        });
+      });
+    });
+
+    it('should return error when setting properties with static website disabled', function (done) {
+      var prop = {
+        StaticWebsite: {
+          IndexDocument: 'index.htm',
+          ErrorDocument404Path: 'error/index.htm'
+        }
+      };
+
+      blobService.setServiceProperties(prop, function (err, res) {
+        assert.notEqual(err, null);
+        done();
+      });
+    });
   });
 });