+++ typings/azure-storage/azure-storage.d.ts
@@ -159,8 +159,8 @@ declare module azurestorage {
           * @param {errorOrResult}  callback                              `error` will contain information if an error occurs; otherwise, `result` will contain the properties
           *                                                               and `response` will contain information related to this operation.
           */
-          getServiceProperties(options: common.RequestOptions, callback?: ErrorOrResult<common.models.ServicePropertiesResult.ServiceProperties>): void;
-          getServiceProperties(callback?: ErrorOrResult<common.models.ServicePropertiesResult.ServiceProperties>): void;
+          getServiceProperties(options: common.RequestOptions, callback?: ErrorOrResult<common.models.ServicePropertiesResult.BlobServiceProperties>): void;
+          getServiceProperties(callback?: ErrorOrResult<common.models.ServicePropertiesResult.BlobServiceProperties>): void;
           /**
           * Sets the properties of a storage account’s Blob service, including Azure Storage Analytics.
           * You can also use this operation to set the default request version for all incoming requests that do not have a version specified.
@@ -182,8 +182,8 @@ declare module azurestorage {
           *                                                                      if an error occurs; otherwise, `response`
           *                                                                      will contain information related to this operation.
           */
-          setServiceProperties(serviceProperties: common.models.ServicePropertiesResult.ServiceProperties, options: common.RequestOptions, callback: ErrorOrResponse): void;
-          setServiceProperties(serviceProperties: common.models.ServicePropertiesResult.ServiceProperties, callback: ErrorOrResponse): void;
+          setServiceProperties(serviceProperties: common.models.ServicePropertiesResult.BlobServiceProperties, options: common.RequestOptions, callback: ErrorOrResponse): void;
+          setServiceProperties(serviceProperties: common.models.ServicePropertiesResult.BlobServiceProperties, callback: ErrorOrResponse): void;
 
           /**
           * Sets the tier of a blockblob under a blob storage LRS account, or the tier of a pageblob under a premium storage account.
@@ -9080,6 +9080,14 @@ declare module azurestorage {
           Cors?: {
             CorsRule: CorsRule[];
           };
+        }
+        export interface StaticWebsiteProperties {
+          Enabled: boolean;
+          IndexDocument?: string;
+          ErrorDocument404Path?: string;
+        }
+        export interface BlobServiceProperties extends ServiceProperties {
+          StaticWebsite?: StaticWebsiteProperties;
         }
         export function serialize(servicePropertiesJs: ServiceProperties): string;
         export function parse(servicePropertiesXml: any): ServiceProperties;